# Decoupled API structs from efficient scheduling representation and deprecated constraints object