--- a/pkg/apis/awsnodetemplate/v1alpha1/register.go
+++ b/pkg/apis/awsnodetemplate/v1alpha1/register.go
@@ -23,7 +23,6 @@ import (
 
 var (
 	Group              = "karpenter.k8s.aws"
-	ExtensionsGroup    = "extensions." + Group
 	SchemeGroupVersion = schema.GroupVersion{Group: Group, Version: "v1alpha1"}
 	SchemeBuilder      = runtime.NewSchemeBuilder(func(scheme *runtime.Scheme) error {
 		scheme.AddKnownTypes(SchemeGroupVersion,