--- a/pkg/cloudprovider/aws/cloudprovider.go
+++ b/pkg/cloudprovider/aws/cloudprovider.go
@@ -32,12 +32,13 @@ import (
 	"github.com/aws/karpenter/pkg/cloudprovider"
 	"github.com/aws/karpenter/pkg/cloudprovider/aws/amifamily"
 	"github.com/aws/karpenter/pkg/cloudprovider/aws/apis/v1alpha1"
+	"github.com/aws/karpenter/pkg/scheduling"
 	"github.com/aws/karpenter/pkg/utils/functional"
 	"github.com/aws/karpenter/pkg/utils/injection"
 	"github.com/aws/karpenter/pkg/utils/project"
+	"github.com/aws/karpenter/pkg/utils/sets"
 
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/transport"
 	"knative.dev/pkg/apis"
 	"knative.dev/pkg/logging"
@@ -118,23 +119,21 @@ func (c *CloudProvider) Delete(ctx context.Context, node *v1.Node) error {
 	return c.instanceProvider.Terminate(ctx, node)
 }
 
-func (c *CloudProvider) GetRequirements(ctx context.Context, provider *v1alpha5.Provider) (v1alpha5.Requirements, error) {
+func (c *CloudProvider) GetRequirements(ctx context.Context, provider *v1alpha5.Provider) (scheduling.Requirements, error) {
 	awsprovider, err := v1alpha1.Deserialize(provider)
 	if err != nil {
-		return v1alpha5.Requirements{}, apis.ErrGeneric(err.Error())
+		return scheduling.NewRequirements(), apis.ErrGeneric(err.Error())
 	}
 	// Constrain AZs from subnets
 	subnets, err := c.subnetProvider.Get(ctx, awsprovider)
 	if err != nil {
-		return v1alpha5.Requirements{}, err
+		return scheduling.NewRequirements(), err
 	}
-	zones := sets.NewString()
+	zones := sets.NewSet()
 	for _, subnet := range subnets {
 		zones.Insert(aws.StringValue(subnet.AvailabilityZone))
 	}
-	requirements := v1alpha5.NewRequirements(v1.NodeSelectorRequirement{
-		Key: v1.LabelTopologyZone, Operator: v1.NodeSelectorOpIn, Values: zones.List(),
-	})
+	requirements := scheduling.Requirements{v1.LabelTopologyZone: zones}
 	return requirements, nil
 }
 
@@ -161,13 +160,13 @@ func defaultLabels(provisioner *v1alpha5.Provisioner) {
 		if _, ok := provisioner.Spec.Labels[key]; ok {
 			hasLabel = true
 		}
-		for _, requirement := range provisioner.Spec.Requirements.Requirements {
+		for _, requirement := range provisioner.Spec.Requirements {
 			if requirement.Key == key {
 				hasLabel = true
 			}
 		}
 		if !hasLabel {
-			provisioner.Spec.Requirements.Requirements = append(provisioner.Spec.Requirements.Requirements, v1.NodeSelectorRequirement{
+			provisioner.Spec.Requirements = append(provisioner.Spec.Requirements, v1.NodeSelectorRequirement{
 				Key: key, Operator: v1.NodeSelectorOpIn, Values: []string{value},
 			})
 		}