--- a/pkg/cloudprovider/fake/cloudprovider.go
+++ b/pkg/cloudprovider/fake/cloudprovider.go
@@ -28,6 +28,7 @@ import (
 	"knative.dev/pkg/apis"
 
 	"github.com/aws/karpenter/pkg/cloudprovider/aws/apis/v1alpha1"
+	"github.com/aws/karpenter/pkg/scheduling"
 
 	"github.com/aws/karpenter/pkg/apis/provisioning/v1alpha5"
 	"github.com/aws/karpenter/pkg/cloudprovider"
@@ -45,12 +46,6 @@ type CloudProvider struct {
 	// CreateCalls contains the arguments for every create call that was made since it was cleared
 	mu          sync.Mutex
 	CreateCalls []*cloudprovider.NodeRequest
-}
-
-type CreateCallArgs struct {
-	Constraints   *v1alpha5.Constraints
-	InstanceTypes []cloudprovider.InstanceType
-	Quantity      int
 }
 
 func (c *CloudProvider) Create(ctx context.Context, nodeRequest *cloudprovider.NodeRequest) (*v1.Node, error) {
@@ -143,8 +138,8 @@ func (c *CloudProvider) GetInstanceTypes(_ context.Context) ([]cloudprovider.Ins
 	}, nil
 }
 
-func (c *CloudProvider) GetRequirements(ctx context.Context, provider *v1alpha5.Provider) (v1alpha5.Requirements, error) {
-	return v1alpha5.Requirements{}, nil
+func (c *CloudProvider) GetRequirements(ctx context.Context, provider *v1alpha5.Provider) (scheduling.Requirements, error) {
+	return scheduling.NewRequirements(), nil
 }
 
 func (c *CloudProvider) Delete(context.Context, *v1.Node) error {