--- a/pkg/cloudprovider/types.go
+++ b/pkg/cloudprovider/types.go
@@ -23,6 +23,7 @@ import (
 	"knative.dev/pkg/apis"
 
 	"github.com/aws/karpenter/pkg/apis/provisioning/v1alpha5"
+	"github.com/aws/karpenter/pkg/scheduling"
 )
 
 // Options are injected into cloud providers' factories
@@ -44,7 +45,7 @@ type CloudProvider interface {
 	GetInstanceTypes(context.Context) ([]InstanceType, error)
 	// GetRequirements for the provider, e.g. zones contrained by subnets or
 	// os constrained by machine image.
-	GetRequirements(context.Context, *v1alpha5.Provider) (v1alpha5.Requirements, error)
+	GetRequirements(context.Context, *v1alpha5.Provider) (scheduling.Requirements, error)
 	// Default is a hook for additional defaulting logic at webhook time.
 	Default(context.Context, *v1alpha5.Provisioner)
 	// Validate is a hook for additional validation logic at webhook time.
@@ -54,16 +55,8 @@ type CloudProvider interface {
 }
 
 type NodeRequest struct {
-	Template            *NodeTemplate
+	Template            *scheduling.NodeTemplate
 	InstanceTypeOptions []InstanceType
-}
-
-type NodeTemplate struct {
-	Provider             *v1alpha5.Provider
-	Labels               map[string]string
-	Taints               []v1.Taint
-	Requirements         v1alpha5.Requirements
-	KubeletConfiguration *v1alpha5.KubeletConfiguration
 }
 
 // InstanceType describes the properties of a potential node (either concrete attributes of an instance of this type