**1. Issue, if available:**
https://github.com/aws/karpenter/issues/1350

**2. Description of changes:**
Changes:
- New structs at `pkg/scheduling`. 
- We translate from the API to these, and then speak in terms of them for the rest of the codebase. 
- `Add` is now non-chainable but mutates the struct. This improves scheduling performance. 

Before (`8b660ce3`: last commit before benchmark broke)
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkScheduling1-12       	     589	   2229986 ns/op	         1.000 nodes	         1.000 pods	       448.4 pods/sec
BenchmarkScheduling50-12      	      44	  28443412 ns/op	         1.000 nodes	        50.00 pods	      1758 pods/sec
BenchmarkScheduling100-12     	      19	  53785810 ns/op	         1.000 nodes	       100.0 pods	      1859 pods/sec
BenchmarkScheduling500-12     	       6	 174541013 ns/op	         1.000 nodes	       500.0 pods	      2865 pods/sec
BenchmarkScheduling1000-12    	       3	 464866274 ns/op	         3.000 nodes	      1000 pods	      2151 pods/sec
BenchmarkScheduling2000-12    	       1	1107541779 ns/op	         5.000 nodes	      2000 pods	      1806 pods/sec
BenchmarkScheduling5000-12    	       1	5731174687 ns/op	        12.00 nodes	      5000 pods	       872.4 pods/sec
```

After 
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkScheduling1-12       	    4378	    267545 ns/op	         1.000 nodes	         1.000 pods	      3738 pods/sec
BenchmarkScheduling50-12      	      84	  14235042 ns/op	         1.000 nodes	        50.00 pods	      3512 pods/sec
BenchmarkScheduling100-12     	      43	  27653722 ns/op	         1.000 nodes	       100.0 pods	      3616 pods/sec
BenchmarkScheduling500-12     	       7	 153111794 ns/op	         1.000 nodes	       500.0 pods	      3266 pods/sec
BenchmarkScheduling1000-12    	       3	 385912960 ns/op	         3.000 nodes	      1000 pods	      2591 pods/sec
BenchmarkScheduling2000-12    	       2	 932610822 ns/op	         5.000 nodes	      2000 pods	      2145 pods/sec
BenchmarkScheduling5000-12    	       1	3345053020 ns/op	        12.00 nodes	      5000 pods	      1495 pods/sec
```


**3. How was this change tested?**


**4. Does this change impact docs?**
- [ ] Yes, PR includes docs updates
- [ ] Yes, issue opened: *link to issue*
- [ ] No

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.